+++ internal/ci/runner.go
@@ -5,10 +5,10 @@ import (
 	"context"
 	"fmt"
 	"github.com/saucelabs/saucectl/cli/runner"
+	"github.com/saucelabs/saucectl/cli/utils"
 	"github.com/saucelabs/saucectl/internal/fleet"
 	"github.com/saucelabs/saucectl/internal/fpath"
 	"github.com/saucelabs/saucectl/internal/yaml"
-	"github.com/saucelabs/saucectl/internal/utils"
 	"io/ioutil"
 	"os"
 	"os/exec"
