+++ internal/docker/runner.go
@@ -6,9 +6,9 @@ import (
 	"github.com/rs/zerolog/log"
 	"github.com/saucelabs/saucectl/cli/runner"
 	"github.com/saucelabs/saucectl/cli/streams"
+	"github.com/saucelabs/saucectl/cli/utils"
 	"github.com/saucelabs/saucectl/internal/fleet"
 	"github.com/saucelabs/saucectl/internal/yaml"
-	"github.com/saucelabs/saucectl/internal/utils"
 	"io"
 	"io/ioutil"
 	"os"
